Fix to return single latest record for Bar and Instrument from Parquet Catalog (#839)
This is the fix I have made for duplicate records received for dataframe/nautilus objects and part is reported in #833 . This makes sure that the returned records are the latest single record being returned. Also subset keyword is added to allow custom unique filtering for any other classes not added above.