Fix potential iteration of None linked_order_ids
Improve robustness of matching engine iterations where we don't expect
the `order.linked_order_ids` to be `None`, but for some reason based
on upstream logic the `linked_order_ids` is `None`.

We suffer the small performance cost of allocating an empty python list
to avoid an unhandled `TypeError` at run time.